requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
pandas==1.4.2
waitress==2.0.0
whitenoise==5.2.0
